db migrations for vectors
