+++ supabase/functions/_shared/types/constants.ts
@@ -1,0 +1,3 @@
+export const SUPABASE_URL = Deno.env.get("LOCAL_SUPABASE_URL") || Deno.env.get("SUPABASE_URL") || "";
+export const SUPABASE_ANON_KEY = Deno.env.get("LOCAL_SUPABASE_ANON_KEY") || Deno.env.get("SUPABASE_ANON_KEY") || "";
+
